requirements.txt
asgiref==3.8.1
Django==4.2.2
django-phonenumber-field==8.1.0
phonenumbers==9.0.3
sqlparse==0.5.3
